:zap:[#14] 함수 viewModel로 분리
